+++ MyApplication/app/src/main/java/com/skypan/myapplication/driver_model/addSetting.java
@@ -1,6 +1,6 @@
-package com.skypan.myapplication;
+package com.skypan.myapplication.driver_model;
 
-import com.skypan.myapplication.ui.Setting;
+import com.skypan.myapplication.driver_model.ui.Setting;
 
 import java.util.ArrayList;
