+++ MyApplication/app/src/main/java/com/skypan/myapplication/driver_model/ui/Setting.java
@@ -1,4 +1,4 @@
-package com.skypan.myapplication.ui;
+package com.skypan.myapplication.driver_model.ui;
 
 
 import android.text.Editable;
