+++ MyApplication/app/src/main/java/com/skypan/myapplication/driver_model/ui/Switch/SwitchFragment.java
@@ -1,4 +1,4 @@
-package com.skypan.myapplication.ui.Switch;
+package com.skypan.myapplication.driver_model.ui.Switch;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
