+++ MyApplication/app/src/main/java/com/skypan/myapplication/driver_model/ui/history/HistoryViewModel.java
@@ -1,16 +1,16 @@
-package com.skypan.myapplication.ui.home;
+package com.skypan.myapplication.driver_model.ui.history;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class HomeViewModel extends ViewModel {
+public class HistoryViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public HomeViewModel() {
+    public HistoryViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is home fragment");
+        mText.setValue("This is slideshow fragment");
     }
 
     public LiveData<String> getText() {
