+++ MyApplication/app/src/main/java/com/skypan/myapplication/driver_model/ui/home/HomeViewModel.java
@@ -1,16 +1,16 @@
-package com.skypan.myapplication.ui.history;
+package com.skypan.myapplication.driver_model.ui.home;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class HistoryViewModel extends ViewModel {
+public class HomeViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public HistoryViewModel() {
+    public HomeViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is slideshow fragment");
+        mText.setValue("This is home fragment");
     }
 
     public LiveData<String> getText() {
