+++ MyApplication/app/src/main/java/com/skypan/myapplication/login_model/set_new_passwordActivity.java
@@ -1,12 +1,13 @@
-package com.skypan.myapplication;
+package com.skypan.myapplication.login_model;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.TextView;
+
+import com.skypan.myapplication.R;
 
 public class set_new_passwordActivity extends AppCompatActivity {
     private TextView cancel_new_password;
