+++ MyApplication/app/src/main/java/com/skypan/myapplication/login_model/verification_forgetActivity.java
@@ -1,10 +1,12 @@
-package com.skypan.myapplication;
+package com.skypan.myapplication.login_model;
 
 import androidx.appcompat.app.AppCompatActivity;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.TextView;
+
+import com.skypan.myapplication.R;
 
 public class verification_forgetActivity extends AppCompatActivity {
     private TextView cancel_verification_forget;
