+++ MyApplication/app/src/main/res/layout/activity_forget_password.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".forget_password">
+    tools:context=".login_model.forget_password">
 
     <!--    這是 forget_password 介面    -->
     <LinearLayout
