+++ MyApplication/app/src/main/res/layout/activity_profile.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.skypan.myapplication.ui.logout.">
+    tools:context=".driver_model.ui.logout.">
 
     <TextView
         android:id="@+id/tv_1"
