+++ MyApplication/app/src/main/res/layout/activity_text.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@android:color/darker_gray"
-    tools:context=".DriverMainActivity"
+    tools:context=".driver_model.DriverMainActivity"
     android:orientation="vertical">
 
     <Button
