+++ MyApplication/app/src/main/res/layout/fragment_history.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.history.HistoryFragment">
+    tools:context=".driver_model.ui.history.HistoryFragment">
 
     <TextView
         android:id="@+id/text_history"
