+++ MyApplication/app/src/main/res/layout/fragment_logout.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.logout.LogoutFragment">
+    tools:context=".driver_model.ui.logout.LogoutFragment">
 
     <TextView
         android:id="@+id/text_logout"
