+++ MyApplication/app/src/main/res/layout/fragment_switch.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.Switch.SwitchFragment">
+    tools:context=".driver_model.ui.Switch.SwitchFragment">
 
     <TextView
         android:id="@+id/text_switch"
